Changes after merge conflicts with platform e2e tests
